+++ CHANGES.md
@@ -1,6 +1,8 @@
-# (Unreleased)
-- [FIXED] An issue where `Changes.hasNext()` never returns on receipt
-  of `last_seq` for continuous changes feeds.
+# UNRELEASED
+- [NEW] Added an extra bluemix method to the client builder allowing a custom service name to be
+  used with the VCAP_SERVICES environment variable content.
+- [FIXED] An issue where `Changes.hasNext()` never returns on receipt of `last_seq` for continuous
+  changes feeds.
 
 # 2.10.0 (2017-11-07)
 - [NEW] Added API for upcoming IBM Cloud Identity and Access Management support for Cloudant on IBM